requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.53
langchain-groq==0.1.3

chromadb==0.4.22

sentence-transformers==2.6.1
huggingface-hub==0.22.2

ragas==0.1.9

python-dotenv==1.0.1
pydantic==2.6.4
tqdm==4.66.2
